+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 1.0.1-Beta
+- Add Boundless Exploration Advancement
 - Add Sweet Dreams Music Disc as Reward for completing Ultimate Adventurer
 - Add Ambient Sound Subtitle